123/requirement.txt
2captcha-python==1.5.1
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
cloudscraper==1.2.71
fake-useragent==2.2.0
greenlet==3.2.4
idna==3.10
playwright==1.54.0
pyaes==1.6.1
pyee==13.0.0
PyJWT==2.10.1
pyparsing==3.2.3
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
requests-toolbelt==1.0.0
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
Flask>=2.0.1
